Make Start Logic
플레이어의 언어, 이름, 성별, 생일을 Player 스크립트에 저장

플레이어의 정보 세팅이 끝나면 프롤로그 대화 출력
